requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
anyio==3.2.1
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
ci-info==0.2.0
click==8.0.1
configparser==5.0.2
cycler==0.10.0
debugpy==1.3.0
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.12
docker-pycreds==0.4.0
einops==0.3.2
entrypoints==0.3
et-xmlfile==1.1.0
etelemetry==0.2.2
filelock==3.0.12
fpdf==1.7.2
fsspec==2021.6.1
future==0.18.2
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.32.1
google-auth-oauthlib==0.4.4
graphql-core==3.1.5
grpcio==1.38.1
humanize==3.9.0
idna==2.10
imageio==2.9.0
ipykernel==6.0.0
ipython==7.25.0
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.9.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
monai==0.8.0
multidict==5.1.0
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nibabel==3.2.1
nipype==1.6.1
notebook==6.4.0
numpy
oauthlib==3.1.1
opencv-python==4.5.3.56
openpyxl==3.0.9
packaging==20.9
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
pathtools==0.1.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pkg_resources==0.0.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.19
protobuf==3.17.3
prov==2.0.0
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyDeprecate==0.3.1
pydicom==2.1.2
pydot==1.4.2
Pygments==2.9.0
pyminc==0.54
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
pytorch-lightning==1.4.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
rdflib==5.0.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
-e git+https://github.com/CAAI/rh-scripts.git@f316876afdc434107d8f442d93088aa07da6cff0#egg=rhscripts
git+https://github.com/CAAI/rh-torch@5fa8de530e21345d987bb3988890c65daae052c0
rsa==4.7.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
Send2Trash==1.7.1
sentry-sdk==1.1.0
shortuuid==1.0.1
SimpleITK==1.2.4
simplejson==3.17.2
six==1.16.0
smmap==4.0.0
sniffio==1.2.0
statsmodels==0.13.0
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.1.0
tifffile==2021.7.2
timm==0.4.12
torch==1.9.0+cu111
torch-dct==0.1.5
torchaudio==0.9.0
torchio==0.18.46
torchmetrics==0.4.1
torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
traits==6.2.0
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
wrapt==1.12.1
xlrd==2.0.1
yarl==1.6.3
yaspin==2.1.0
